Renaming Service -> OldService, NewService -> Service
