+++ eclair-core/src/main/scala/fr/acinq/eclair/api/Service.scala
@@ -1,367 +1,66 @@
-/*
- * Copyright 2018 ACINQ SAS
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package fr.acinq.eclair.api
 
-import akka.NotUsed
-import akka.actor.{Actor, ActorRef, ActorSystem, Props, Scheduler}
-import akka.http.scaladsl.model.HttpMethods._
-import akka.http.scaladsl.model._
-import akka.http.scaladsl.model.headers.CacheDirectives.{`max-age`, `no-store`, public}
-import akka.http.scaladsl.model.headers._
-import akka.http.scaladsl.model.ws.{Message, TextMessage}
-import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server._
-import akka.http.scaladsl.server.directives.Credentials
-import akka.http.scaladsl.server.directives.RouteDirectives.reject
-import akka.pattern.ask
-import akka.stream.scaladsl.{BroadcastHub, Flow, Keep, Source}
-import akka.stream.{ActorMaterializer, OverflowStrategy}
-import akka.util.Timeout
-import de.heikoseeberger.akkahttpjson4s.Json4sSupport
-import de.heikoseeberger.akkahttpjson4s.Json4sSupport.ShouldWritePretty
 import fr.acinq.bitcoin.Crypto.PublicKey
 import fr.acinq.bitcoin.{ByteVector32, MilliSatoshi, Satoshi}
-import fr.acinq.eclair.channel._
-import fr.acinq.eclair.io.Peer.{GetPeerInfo, PeerInfo}
-import fr.acinq.eclair.io.{NodeURI, Peer}
-import fr.acinq.eclair.payment.PaymentLifecycle._
-import fr.acinq.eclair.payment._
-import fr.acinq.eclair.router.{ChannelDesc, RouteRequest, RouteResponse}
-import fr.acinq.eclair.wire.{ChannelAnnouncement, ChannelUpdate, NodeAddress, NodeAnnouncement}
-import fr.acinq.eclair.{Kit, ShortChannelId, feerateByte2Kw}
+import fr.acinq.eclair.{Kit, ShortChannelId}
+import FormParamExtractors._
+import akka.NotUsed
+import akka.actor.{Actor, ActorRef, ActorSystem, Props}
+import akka.http.scaladsl.model.HttpMethods.POST
+import akka.http.scaladsl.model.{ContentTypes, HttpRequest, HttpResponse, StatusCodes}
+import akka.http.scaladsl.model.headers.CacheDirectives.{`max-age`, `no-store`, public}
+import akka.http.scaladsl.model.headers.{`Access-Control-Allow-Headers`, `Access-Control-Allow-Methods`, `Cache-Control`}
+import akka.http.scaladsl.model.ws.{Message, TextMessage}
+import akka.http.scaladsl.server.directives.{Credentials, LoggingMagnet}
+import akka.stream.{ActorMaterializer, OverflowStrategy}
+import akka.stream.scaladsl.{BroadcastHub, Flow, Keep, Source}
+import fr.acinq.eclair.payment.{PaymentLifecycle, PaymentReceived, PaymentRequest}
 import grizzled.slf4j.Logging
-import org.json4s.JsonAST.{JBool, JInt, JString}
-import org.json4s.{JValue, jackson}
 import scodec.bits.ByteVector
-
-import scala.concurrent.duration._
 import scala.concurrent.{ExecutionContext, Future}
-import scala.util.{Failure, Success, Try}
+import scala.concurrent.duration._
 
-// @formatter:off
-case class JsonRPCBody(jsonrpc: String = "1.0", id: String = "eclair-node", method: String, params: Seq[JValue])
-case class Error(code: Int, message: String)
-case class JsonRPCRes(result: AnyRef, error: Option[Error], id: String)
-case class Status(node_id: String)
-case class GetInfoResponse(nodeId: PublicKey, alias: String, port: Int, chainHash: ByteVector32, blockHeight: Int, publicAddresses: Seq[NodeAddress])
-case class AuditResponse(sent: Seq[PaymentSent], received: Seq[PaymentReceived], relayed: Seq[PaymentRelayed])
-trait RPCRejection extends Rejection {
-  def requestId: String
-}
-final case class UnknownMethodRejection(requestId: String) extends RPCRejection
-final case class UnknownParamsRejection(requestId: String, message: String) extends RPCRejection
-final case class RpcValidationRejection(requestId: String, message: String) extends RPCRejection
-final case class ExceptionRejection(requestId: String, message: String) extends RPCRejection
-// @formatter:on
+trait Service extends Directives with Logging with MetaService {
 
-trait Service extends Logging with MetaService {
-
-  implicit def ec: ExecutionContext = ExecutionContext.Implicits.global
-
-  def scheduler: Scheduler
-
-  implicit val serialization = jackson.Serialization
-  implicit val formats = org.json4s.DefaultFormats + new ByteVectorSerializer + new ByteVector32Serializer + new UInt64Serializer + new MilliSatoshiSerializer + new ShortChannelIdSerializer + new StateSerializer + new ShaChainSerializer + new PublicKeySerializer + new PrivateKeySerializer + new ScalarSerializer + new PointSerializer + new TransactionSerializer + new TransactionWithInputInfoSerializer + new InetSocketAddressSerializer + new OutPointSerializer + new OutPointKeySerializer + new InputInfoSerializer + new ColorSerializer +  new RouteResponseSerializer + new ThrowableSerializer + new FailureMessageSerializer + new NodeAddressSerializer + new DirectionSerializer +new PaymentRequestSerializer
-  implicit val timeout = Timeout(60 seconds)
-  implicit val shouldWritePretty: ShouldWritePretty = ShouldWritePretty.True
-
-  import Json4sSupport.{marshaller, unmarshaller}
+  // important! Must NOT import the unmarshaller as it is too generic...see https://github.com/akka/akka-http/issues/541
+  import JsonSupport.marshaller
+  import JsonSupport.formats
+  import JsonSupport.serialization
 
   def password: String
 
-  def appKit: Kit
+  def eclairApi: EclairApi
 
-  val socketHandler: Flow[Message, TextMessage.Strict, NotUsed]
+  implicit val actorSystem: ActorSystem
+  implicit val mat: ActorMaterializer
 
-  def userPassAuthenticator(credentials: Credentials): Future[Option[String]] = credentials match {
-    case p@Credentials.Provided(id) if p.verify(password) => Future.successful(Some(id))
-    case _ => akka.pattern.after(1 second, using = scheduler)(Future.successful(None)) // force a 1 sec pause to deter brute force
+  // a named and typed URL parameter used across several routes, 32-bytes hex-encoded
+  val channelIdNamedParameter = "channelId".as[ByteVector32](sha256HashUnmarshaller)
+  val shortChannelIdNamedParameter = "shortChannelId".as[ShortChannelId](shortChannelIdUnmarshaller)
+
+  val apiExceptionHandler = ExceptionHandler {
+    case t: Throwable =>
+      logger.error(s"API call failed with cause=${t.getMessage}", t)
+      complete(StatusCodes.InternalServerError, s"Error: $t")
   }
 
   val customHeaders = `Access-Control-Allow-Headers`("Content-Type, Authorization") ::
     `Access-Control-Allow-Methods`(POST) ::
     `Cache-Control`(public, `no-store`, `max-age`(0)) :: Nil
 
-  val myExceptionHandler = ExceptionHandler {
-    case t: Throwable =>
-      extractRequest { _ =>
-        logger.error(s"API call failed with cause=${t.getMessage}")
-        complete(StatusCodes.InternalServerError, JsonRPCRes(null, Some(Error(StatusCodes.InternalServerError.intValue, t.getMessage)), "-1"))
-      }
-  }
-
-  def completeRpcFuture(requestId: String, future: Future[AnyRef]): Route = onComplete(future) {
-    case Success(s) => completeRpc(requestId, s)
-    case Failure(t) => reject(ExceptionRejection(requestId, t.getLocalizedMessage))
-  }
-
-  def completeRpc(requestId: String, result: AnyRef): Route = complete(JsonRPCRes(result, None, requestId))
-
-  val myRejectionHandler: RejectionHandler = RejectionHandler.newBuilder()
-    .handleNotFound {
-      complete(StatusCodes.NotFound, JsonRPCRes(null, Some(Error(StatusCodes.NotFound.intValue, "not found")), "-1"))
-    }
-    .handle {
-      case _: AuthenticationFailedRejection ⇒ complete(StatusCodes.Unauthorized, JsonRPCRes(null, Some(Error(StatusCodes.Unauthorized.intValue, "Access restricted")), "-1"))
-      case v: RpcValidationRejection ⇒ complete(StatusCodes.BadRequest, JsonRPCRes(null, Some(Error(StatusCodes.BadRequest.intValue, v.message)), v.requestId))
-      case ukm: UnknownMethodRejection ⇒ complete(StatusCodes.BadRequest, JsonRPCRes(null, Some(Error(StatusCodes.BadRequest.intValue, "method not found")), ukm.requestId))
-      case p: UnknownParamsRejection ⇒ complete(StatusCodes.BadRequest,
-        JsonRPCRes(null, Some(Error(StatusCodes.BadRequest.intValue, s"invalid parameters for this method, should be: ${p.message}")), p.requestId))
-      case m: MalformedRequestContentRejection ⇒ complete(StatusCodes.BadRequest,
-        JsonRPCRes(null, Some(Error(StatusCodes.BadRequest.intValue, s"malformed parameters for this method: ${m.message}")), "-1"))
-      case e: ExceptionRejection ⇒ complete(StatusCodes.BadRequest,
-        JsonRPCRes(null, Some(Error(StatusCodes.BadRequest.intValue, s"command failed: ${e.message}")), e.requestId))
-      case r ⇒ logger.error(s"API call failed with cause=$r")
-        complete(StatusCodes.BadRequest, JsonRPCRes(null, Some(Error(StatusCodes.BadRequest.intValue, r.toString)), "-1"))
-    }
-    .result()
-
-  val route: Route =
-    respondWithDefaultHeaders(customHeaders) {
-      withRequestTimeoutResponse(r => HttpResponse(StatusCodes.RequestTimeout).withEntity(ContentTypes.`application/json`, """{ "result": null, "error": { "code": 408, "message": "request timed out"} } """)) {
-        handleExceptions(myExceptionHandler) {
-          handleRejections(myRejectionHandler) {
-            authenticateBasicAsync(realm = "Access restricted", userPassAuthenticator) { _ =>
-              pathSingleSlash {
-                post {
-                  entity(as[JsonRPCBody]) {
-                    req =>
-                      val kit = appKit
-                      import kit._
-
-                      req.method match {
-                        // utility methods
-                        case "getinfo" => completeRpcFuture(req.id, getInfoResponse)
-                        case "help" => completeRpc(req.id, help)
-
-                        // channel lifecycle methods
-                        case "connect" => req.params match {
-                          case JString(pubkey) :: JString(host) :: JInt(port) :: Nil =>
-                            completeRpcFuture(req.id, (switchboard ? Peer.Connect(NodeURI.parse(s"$pubkey@$host:$port"))).mapTo[String])
-                          case JString(uri) :: Nil =>
-                            completeRpcFuture(req.id, (switchboard ? Peer.Connect(NodeURI.parse(uri))).mapTo[String])
-                          case _ => reject(UnknownParamsRejection(req.id, "[nodeId@host:port] or [nodeId, host, port]"))
-                        }
-                        case "open" => req.params match {
-                          case JString(nodeId) :: JInt(fundingSatoshis) :: Nil =>
-                            completeRpcFuture(req.id, (switchboard ? Peer.OpenChannel(PublicKey(ByteVector.fromValidHex(nodeId)), Satoshi(fundingSatoshis.toLong), MilliSatoshi(0), fundingTxFeeratePerKw_opt = None, channelFlags = None)).mapTo[String])
-                          case JString(nodeId) :: JInt(fundingSatoshis) :: JInt(pushMsat) :: Nil =>
-                            completeRpcFuture(req.id, (switchboard ? Peer.OpenChannel(PublicKey(ByteVector.fromValidHex(nodeId)), Satoshi(fundingSatoshis.toLong), MilliSatoshi(pushMsat.toLong), channelFlags = None, fundingTxFeeratePerKw_opt = None)).mapTo[String])
-                          case JString(nodeId) :: JInt(fundingSatoshis) :: JInt(pushMsat) :: JInt(fundingFeerateSatPerByte) :: Nil =>
-                            completeRpcFuture(req.id, (switchboard ? Peer.OpenChannel(PublicKey(ByteVector.fromValidHex(nodeId)), Satoshi(fundingSatoshis.toLong), MilliSatoshi(pushMsat.toLong), fundingTxFeeratePerKw_opt = Some(feerateByte2Kw(fundingFeerateSatPerByte.toLong)), channelFlags = None)).mapTo[String])
-                          case JString(nodeId) :: JInt(fundingSatoshis) :: JInt(pushMsat) :: JInt(fundingFeerateSatPerByte) :: JInt(flags) :: Nil =>
-                            completeRpcFuture(req.id, (switchboard ? Peer.OpenChannel(PublicKey(ByteVector.fromValidHex(nodeId)), Satoshi(fundingSatoshis.toLong), MilliSatoshi(pushMsat.toLong), fundingTxFeeratePerKw_opt = Some(feerateByte2Kw(fundingFeerateSatPerByte.toLong)), channelFlags = Some(flags.toByte))).mapTo[String])
-                          case _ => reject(UnknownParamsRejection(req.id, s"[nodeId, fundingSatoshis], [nodeId, fundingSatoshis, pushMsat], [nodeId, fundingSatoshis, pushMsat, feerateSatPerByte] or [nodeId, fundingSatoshis, pushMsat, feerateSatPerByte, flag]"))
-                        }
-                        case "close" => req.params match {
-                          case JString(identifier) :: Nil => completeRpcFuture(req.id, sendToChannel(identifier, CMD_CLOSE(scriptPubKey = None)).mapTo[String])
-                          case JString(identifier) :: JString(scriptPubKey) :: Nil => completeRpcFuture(req.id, sendToChannel(identifier, CMD_CLOSE(scriptPubKey = Some(ByteVector.fromValidHex(scriptPubKey)))).mapTo[String])
-                          case _ => reject(UnknownParamsRejection(req.id, "[channelId] or [channelId, scriptPubKey]"))
-                        }
-                        case "forceclose" => req.params match {
-                          case JString(identifier) :: Nil => completeRpcFuture(req.id, sendToChannel(identifier, CMD_FORCECLOSE).mapTo[String])
-                          case _ => reject(UnknownParamsRejection(req.id, "[channelId]"))
-                        }
-                        case "updaterelayfee" => req.params match {
-                          case JString(identifier) :: JInt(feeBaseMsat) :: JInt(feeProportionalMillionths) :: Nil =>
-                            completeRpcFuture(req.id, sendToChannel(identifier, CMD_UPDATE_RELAY_FEE(feeBaseMsat.toLong, feeProportionalMillionths.toLong)).mapTo[String])
-                          case JString(identifier) :: JString(feeBaseMsat) :: JString(feeProportionalMillionths) :: Nil =>
-                            completeRpcFuture(req.id, sendToChannel(identifier, CMD_UPDATE_RELAY_FEE(feeBaseMsat.toLong, feeProportionalMillionths.toLong)).mapTo[String])
-                          case _ => reject(UnknownParamsRejection(req.id, "[channelId] [feeBaseMsat] [feeProportionalMillionths]"))
-                        }
-                        // local network methods
-                        case "peers" => completeRpcFuture(req.id, for {
-                          peers <- (switchboard ? 'peers).mapTo[Iterable[ActorRef]]
-                          peerinfos <- Future.sequence(peers.map(peer => (peer ? GetPeerInfo).mapTo[PeerInfo]))
-                        } yield peerinfos)
-                        case "channels" => req.params match {
-                          case Nil =>
-                            val f = for {
-                              channels_id <- (register ? 'channels).mapTo[Map[ByteVector32, ActorRef]].map(_.keys)
-                              channels <- Future.sequence(channels_id.map(channel_id => sendToChannel(channel_id.toString(), CMD_GETINFO).mapTo[RES_GETINFO]))
-                            } yield channels
-                            completeRpcFuture(req.id, f)
-                          case JString(remoteNodeId) :: Nil => Try(PublicKey(ByteVector.fromValidHex(remoteNodeId))) match {
-                            case Success(pk) =>
-                              val f = for {
-                                channels_id <- (register ? 'channelsTo).mapTo[Map[ByteVector32, PublicKey]].map(_.filter(_._2 == pk).keys)
-                                channels <- Future.sequence(channels_id.map(channel_id => sendToChannel(channel_id.toString(), CMD_GETINFO).mapTo[RES_GETINFO]))
-                              } yield channels
-                              completeRpcFuture(req.id, f)
-                            case Failure(_) => reject(RpcValidationRejection(req.id, s"invalid remote node id '$remoteNodeId'"))
-                          }
-                          case _ => reject(UnknownParamsRejection(req.id, "no arguments or [remoteNodeId]"))
-                        }
-                        case "channel" => req.params match {
-                          case JString(identifier) :: Nil => completeRpcFuture(req.id, sendToChannel(identifier, CMD_GETINFO).mapTo[RES_GETINFO])
-                          case _ => reject(UnknownParamsRejection(req.id, "[channelId]"))
-                        }
-
-                        // global network methods
-                        case "allnodes" => completeRpcFuture(req.id, (router ? 'nodes).mapTo[Iterable[NodeAnnouncement]])
-                        case "allchannels" => completeRpcFuture(req.id, (router ? 'channels).mapTo[Iterable[ChannelAnnouncement]].map(_.map(c => ChannelDesc(c.shortChannelId, c.nodeId1, c.nodeId2))))
-                        case "allupdates" => req.params match {
-                          case JString(nodeId) :: Nil => Try(PublicKey(ByteVector.fromValidHex(nodeId))) match {
-                            case Success(pk) => completeRpcFuture(req.id, (router ? 'updatesMap).mapTo[Map[ChannelDesc, ChannelUpdate]].map(_.filter(e => e._1.a == pk || e._1.b == pk).values))
-                            case Failure(_) => reject(RpcValidationRejection(req.id, s"invalid remote node id '$nodeId'"))
-                          }
-                          case _ => completeRpcFuture(req.id, (router ? 'updates).mapTo[Iterable[ChannelUpdate]])
-                        }
-
-                      // payment methods
-                      case "receive" => req.params match {
-                        // only the payment description is given: user may want to generate a donation payment request
-                        case JString(description) :: Nil =>
-                          completeRpcFuture(req.id, (paymentHandler ? ReceivePayment(None, description)).mapTo[PaymentRequest].map(PaymentRequest.write))
-                        // the amount is now given with the description
-                        case JInt(amountMsat) :: JString(description) :: Nil =>
-                          completeRpcFuture(req.id, (paymentHandler ? ReceivePayment(Some(MilliSatoshi(amountMsat.toLong)), description)).mapTo[PaymentRequest].map(PaymentRequest.write))
-                        case JInt(amountMsat) :: JString(description) :: JInt(expirySeconds) :: Nil =>
-                          completeRpcFuture(req.id, (paymentHandler ? ReceivePayment(Some(MilliSatoshi(amountMsat.toLong)), description, Some(expirySeconds.toLong))).mapTo[PaymentRequest].map(PaymentRequest.write))
-                        case _ => reject(UnknownParamsRejection(req.id, "[description] or [amount, description] or [amount, description, expiryDuration]"))
-                      }
-
-                      // checkinvoice deprecated.
-                      case "parseinvoice" | "checkinvoice" => req.params match {
-                        case JString(paymentRequest) :: Nil => Try(PaymentRequest.read(paymentRequest)) match {
-                          case Success(pr) => completeRpc(req.id,pr)
-                          case Failure(t) => reject(RpcValidationRejection(req.id, s"invalid payment request ${t.getMessage}"))
-                        }
-                        case _ => reject(UnknownParamsRejection(req.id, "[payment_request]"))
-                      }
-
-                      case "findroute" => req.params match {
-                        case JString(nodeId) :: JInt(amountMsat) :: Nil if nodeId.length() == 66 => Try(PublicKey(ByteVector.fromValidHex(nodeId))) match {
-                          case Success(pk) => completeRpcFuture(req.id, (router ? RouteRequest(appKit.nodeParams.nodeId, pk, amountMsat.toLong)).mapTo[RouteResponse])
-                          case Failure(_) => reject(RpcValidationRejection(req.id, s"invalid nodeId hash '$nodeId'"))
-                        }
-                        case JString(paymentRequest) :: Nil => Try(PaymentRequest.read(paymentRequest)) match {
-                          case Success(PaymentRequest(_, Some(amountMsat), _, nodeId , _, _)) => completeRpcFuture(req.id, (router ? RouteRequest(appKit.nodeParams.nodeId, nodeId, amountMsat.toLong)).mapTo[RouteResponse])
-                          case Success(_) => reject(RpcValidationRejection(req.id, s"payment request is missing amount, please specify it"))
-                          case Failure(t) => reject(RpcValidationRejection(req.id, s"invalid payment request ${t.getLocalizedMessage}"))
-                        }
-                        case JString(paymentRequest) :: JInt(amountMsat) :: Nil => Try(PaymentRequest.read(paymentRequest)) match {
-                          case Success(PaymentRequest(_, None, _, nodeId , _, _)) => completeRpcFuture(req.id, (router ? RouteRequest(appKit.nodeParams.nodeId, nodeId, amountMsat.toLong)).mapTo[RouteResponse])
-                          case Success(_) => reject(RpcValidationRejection(req.id, s"amount was specified both in payment request and api call"))
-                          case Failure(t) => reject(RpcValidationRejection(req.id, s"invalid payment request ${t.getLocalizedMessage}"))
-                        }
-                        case _ => reject(UnknownParamsRejection(req.id, "[payment_request] or [payment_request, amountMsat] or [nodeId, amountMsat]"))
-                      }
-
-                      case "send" => req.params match {
-                        // user manually sets the payment information
-                        case JInt(amountMsat) :: JString(paymentHash) :: JString(nodeId) :: Nil =>
-                          (Try(ByteVector32.fromValidHex(paymentHash)), Try(PublicKey(ByteVector.fromValidHex(nodeId)))) match {
-                            case (Success(ph), Success(pk)) => completeRpcFuture(req.id, (paymentInitiator ?
-                              SendPayment(amountMsat.toLong, ph, pk)).mapTo[PaymentResult].map {
-                              case s: PaymentSucceeded => s
-                              case f: PaymentFailed => f.copy(failures = PaymentLifecycle.transformForUser(f.failures))
-                            })
-                            case (Failure(_), _) => reject(RpcValidationRejection(req.id, s"invalid payment hash '$paymentHash'"))
-                            case _ => reject(RpcValidationRejection(req.id, s"invalid node id '$nodeId'"))
-                          }
-                        // user gives a Lightning payment request
-                        case JString(paymentRequest) :: rest => Try(PaymentRequest.read(paymentRequest)) match {
-                          case Success(pr) =>
-                            // setting the payment amount
-                            val amount_msat: Long = (pr.amount, rest) match {
-                              // optional amount always overrides the amount in the payment request
-                              case (_, JInt(amount_msat_override) :: Nil) => amount_msat_override.toLong
-                              case (Some(amount_msat_pr), _) => amount_msat_pr.amount
-                              case _ => throw new RuntimeException("you must manually specify an amount for this payment request")
-                            }
-                            logger.debug(s"api call for sending payment with amount_msat=$amount_msat")
-                            // optional cltv expiry
-                            val sendPayment = pr.minFinalCltvExpiry match {
-                              case None => SendPayment(amount_msat, pr.paymentHash, pr.nodeId)
-                              case Some(minFinalCltvExpiry) => SendPayment(amount_msat, pr.paymentHash, pr.nodeId, assistedRoutes = Nil, minFinalCltvExpiry)
-                            }
-                            completeRpcFuture(req.id, (paymentInitiator ? sendPayment).mapTo[PaymentResult].map {
-                              case s: PaymentSucceeded => s
-                              case f: PaymentFailed => f.copy(failures = PaymentLifecycle.transformForUser(f.failures))
-                            })
-                          case _ => reject(RpcValidationRejection(req.id, s"payment request is not valid"))
-                        }
-                        case _ => reject(UnknownParamsRejection(req.id, "[amountMsat, paymentHash, nodeId or [paymentRequest] or [paymentRequest, amountMsat]"))
-                      }
-
-                        // check received payments
-                        case "checkpayment" => req.params match {
-                          case JString(identifier) :: Nil => completeRpcFuture(req.id, for {
-                            paymentHash <- Try(PaymentRequest.read(identifier)) match {
-                              case Success(pr) => Future.successful(pr.paymentHash)
-                              case _ => Try(ByteVector.fromValidHex(identifier)) match {
-                                case Success(s) => Future.successful(s)
-                                case _ => Future.failed(new IllegalArgumentException("payment identifier must be a payment request or a payment hash"))
-                              }
-                            }
-                            found <- (paymentHandler ? CheckPayment(ByteVector32.fromValidHex(identifier))).map(found => new JBool(found.asInstanceOf[Boolean]))
-                          } yield found)
-                          case _ => reject(UnknownParamsRejection(req.id, "[paymentHash] or [paymentRequest]"))
-                        }
-
-                        // retrieve audit events
-                        case "audit" =>
-                          val (from, to) = req.params match {
-                            case JInt(from) :: JInt(to) :: Nil => (from.toLong, to.toLong)
-                            case _ => (0L, Long.MaxValue)
-                          }
-                          completeRpcFuture(req.id, Future(AuditResponse(
-                            sent = nodeParams.auditDb.listSent(from, to),
-                            received = nodeParams.auditDb.listReceived(from, to),
-                            relayed = nodeParams.auditDb.listRelayed(from, to))
-                          ))
-
-                        case "networkfees" =>
-                          val (from, to) = req.params match {
-                            case JInt(from) :: JInt(to) :: Nil => (from.toLong, to.toLong)
-                            case _ => (0L, Long.MaxValue)
-                          }
-                          completeRpcFuture(req.id, Future(nodeParams.auditDb.listNetworkFees(from, to)))
-
-                        // retrieve fee stats
-                        case "channelstats" => completeRpcFuture(req.id, Future(nodeParams.auditDb.stats))
-
-
-                        // method name was not found
-                        case _ => reject(UnknownMethodRejection(req.id))
-                      }
-                  }
-                }
-              }
-            } ~ path("ws") {
-              handleWebSocketMessages(socketHandler)
-            }
-          }
-        }
-      }
-    }
-
-  def getInfoResponse: Future[GetInfoResponse]
-
-  def makeSocketHandler(system: ActorSystem)(implicit materializer: ActorMaterializer): Flow[Message, TextMessage.Strict, NotUsed] = {
+  lazy val makeSocketHandler: Flow[Message, TextMessage.Strict, NotUsed] = {
 
     // create a flow transforming a queue of string -> string
     val (flowInput, flowOutput) = Source.queue[String](10, OverflowStrategy.dropTail).toMat(BroadcastHub.sink[String])(Keep.both).run()
 
     // register an actor that feeds the queue when a payment is received
-    system.actorOf(Props(new Actor {
+    actorSystem.actorOf(Props(new Actor {
       override def preStart: Unit = context.system.eventStream.subscribe(self, classOf[PaymentReceived])
-      def receive: Receive = { case received: PaymentReceived => flowInput.offer(received.paymentHash.toString) }
+
+      def receive: Receive = {
+        case received: PaymentReceived => flowInput.offer(received.paymentHash.toString)
+      }
     }))
 
     Flow[Message]
@@ -370,7 +69,149 @@ trait Service extends Logging with MetaService {
       .map(TextMessage.apply)
   }
 
-  def help = List(
+  val timeoutResponse: HttpRequest => HttpResponse = { r =>
+    HttpResponse(StatusCodes.RequestTimeout).withEntity(ContentTypes.`application/json`, "request timed out")
+  }
+
+  def userPassAuthenticator(credentials: Credentials): Future[Option[String]] = credentials match {
+    case p@Credentials.Provided(id) if p.verify(password) => Future.successful(Some(id))
+    case _ => akka.pattern.after(1 second, using = actorSystem.scheduler)(Future.successful(None))(actorSystem.dispatcher) // force a 1 sec pause to deter brute force
+  }
+
+  val route: Route = {
+    respondWithDefaultHeaders(customHeaders) {
+      handleExceptions(apiExceptionHandler) {
+        withRequestTimeoutResponse(timeoutResponse) {
+          authenticateBasicAsync(realm = "Access restricted", userPassAuthenticator) { _ =>
+            post {
+              path("getinfo") {
+                complete(eclairApi.getInfoResponse())
+              } ~
+                path("help") {
+                  complete(help)
+                } ~
+                path("connect") {
+                  formFields("uri".as[String]) { uri =>
+                    complete(eclairApi.connect(uri))
+                  } ~ formFields("nodeId".as[PublicKey], "host".as[String], "port".as[Int]) { (nodeId, host, port) =>
+                    complete(eclairApi.connect(s"$nodeId@$host:$port"))
+                  }
+                } ~
+                path("open") {
+                  formFields("nodeId".as[PublicKey], "fundingSatoshis".as[Long], "pushMsat".as[Long].?, "fundingFeerateSatByte".as[Long].?, "channelFlags".as[Int].?) {
+                    (nodeId, fundingSatoshis, pushMsat, fundingFeerateSatByte, channelFlags) =>
+                      complete(eclairApi.open(nodeId, fundingSatoshis, pushMsat, fundingFeerateSatByte, channelFlags))
+                  }
+                } ~
+                path("close") {
+                  formFields(channelIdNamedParameter, "scriptPubKey".as[ByteVector](binaryDataUnmarshaller).?) { (channelId, scriptPubKey_opt) =>
+                    complete(eclairApi.close(Left(channelId), scriptPubKey_opt))
+                  } ~ formFields(shortChannelIdNamedParameter, "scriptPubKey".as[ByteVector](binaryDataUnmarshaller).?) { (shortChannelId, scriptPubKey_opt) =>
+                    complete(eclairApi.close(Right(shortChannelId), scriptPubKey_opt))
+                  }
+                } ~
+                path("forceclose") {
+                  formFields(channelIdNamedParameter) { channelId =>
+                    complete(eclairApi.forceClose(Left(channelId)))
+                  } ~ formFields(shortChannelIdNamedParameter) { shortChannelId =>
+                    complete(eclairApi.forceClose(Right(shortChannelId)))
+                  }
+                } ~
+                path("updaterelayfee") {
+                  formFields(channelIdNamedParameter, "feeBaseMsat".as[Long], "feeProportionalMillionths".as[Long]) { (channelId, feeBase, feeProportional) =>
+                    complete(eclairApi.updateRelayFee(channelId.toString, feeBase, feeProportional))
+                  }
+                } ~
+                path("peers") {
+                  complete(eclairApi.peersInfo())
+                } ~
+                path("channels") {
+                  formFields("toRemoteNodeId".as[PublicKey].?) { toRemoteNodeId_opt =>
+                    complete(eclairApi.channelsInfo(toRemoteNodeId_opt))
+                  }
+                } ~
+                path("channel") {
+                  formFields(channelIdNamedParameter) { channelId =>
+                    complete(eclairApi.channelInfo(channelId))
+                  }
+                } ~
+                path("allnodes") {
+                  complete(eclairApi.allnodes())
+                } ~
+                path("allchannels") {
+                  complete(eclairApi.allchannels())
+                } ~
+                path("allupdates") {
+                  formFields("nodeId".as[PublicKey].?) { nodeId_opt =>
+                    complete(eclairApi.allupdates(nodeId_opt))
+                  }
+                } ~
+                path("receive") {
+                  formFields("description".as[String], "amountMsat".as[Long].?, "expireIn".as[Long].?) { (desc, amountMsat, expire) =>
+                    complete(eclairApi.receive(desc, amountMsat, expire))
+                  }
+                } ~
+                path("parseinvoice") {
+                  formFields("invoice".as[PaymentRequest]) { invoice =>
+                    complete(invoice)
+                  }
+                } ~
+                path("findRouteByInvoice") {
+                  formFields("invoice".as[PaymentRequest], "amountMsat".as[Long].?) {
+                    case (invoice@PaymentRequest(_, Some(amount), _, nodeId, _, _), None) => complete(eclairApi.findRoute(nodeId, amount.toLong, invoice.routingInfo))
+                    case (invoice, Some(overrideAmount)) => complete(eclairApi.findRoute(invoice.nodeId, overrideAmount, invoice.routingInfo))
+                    case _ => reject(MalformedFormFieldRejection("invoice", "The invoice must have an amount or you need to specify one using 'amountMsat'"))
+                  }
+                } ~ path("findRouteByNode") {
+                formFields("nodeId".as[PublicKey], "amountMsat".as[Long]) { (nodeId, amount) =>
+                  complete(eclairApi.findRoute(nodeId, amount))
+                }
+              } ~
+                path("sendToInvoice") {
+                  formFields("invoice".as[PaymentRequest], "amountMsat".as[Long].?) {
+                    case (invoice@PaymentRequest(_, Some(amount), _, nodeId, _, _), None) =>
+                      complete(eclairApi.send(nodeId, amount.toLong, invoice.paymentHash, invoice.routingInfo, invoice.minFinalCltvExpiry))
+                    case (invoice, Some(overrideAmount)) =>
+                      complete(eclairApi.send(invoice.nodeId, overrideAmount, invoice.paymentHash, invoice.routingInfo, invoice.minFinalCltvExpiry))
+                    case _ => reject(MalformedFormFieldRejection("invoice", "The invoice must have an amount or you need to specify one using the field 'amountMsat'"))
+                  }
+                } ~
+                path("sendToNode") {
+                  formFields("amountMsat".as[Long], "paymentHash".as[ByteVector32](sha256HashUnmarshaller), "nodeId".as[PublicKey]) { (amountMsat, paymentHash, nodeId) =>
+                    complete(eclairApi.send(nodeId, amountMsat, paymentHash))
+                  }
+                } ~
+                path("checkpayment") {
+                  formFields("paymentHash".as[ByteVector32](sha256HashUnmarshaller)) { paymentHash =>
+                    complete(eclairApi.checkpayment(paymentHash))
+                  } ~ formFields("invoice".as[PaymentRequest]) { invoice =>
+                    complete(eclairApi.checkpayment(invoice.paymentHash))
+                  }
+                } ~
+                path("audit") {
+                  formFields("from".as[Long].?, "to".as[Long].?) { (from, to) =>
+                    complete(eclairApi.audit(from, to))
+                  }
+                } ~
+                path("networkfees") {
+                  formFields("from".as[Long].?, "to".as[Long].?) { (from, to) =>
+                    complete(eclairApi.networkFees(from, to))
+                  }
+                } ~
+                path("channelstats") {
+                  complete(eclairApi.channelStats())
+                } ~
+                path("ws") {
+                  handleWebSocketMessages(makeSocketHandler)
+                }
+            }
+          }
+        }
+      }
+    }
+  }
+
+  val help = List(
     "connect (uri): open a secure connection to a lightning node",
     "connect (nodeId, host, port): open a secure connection to a lightning node",
     "open (nodeId, fundingSatoshis, pushMsat = 0, feerateSatPerByte = ?, channelFlags = 0x01): open a channel with another lightning node, by default push = 0, feerate for the funding tx targets 6 blocks, and channel is announced",
@@ -405,18 +246,5 @@ trait Service extends Logging with MetaService {
     "getinfo: returns info about the blockchain and this node",
     "help: display this message")
 
-  /**
-    * Sends a request to a channel and expects a response
-    *
-    * @param channelIdentifier can be a shortChannelId (BOLT encoded) or a channelId (32-byte hex encoded)
-    * @param request
-    * @return
-    */
-  def sendToChannel(channelIdentifier: String, request: Any): Future[Any] =
-    for {
-      fwdReq <- Future(Register.ForwardShortId(ShortChannelId(channelIdentifier), request))
-        .recoverWith { case _ => Future(Register.Forward(ByteVector32.fromValidHex(channelIdentifier), request)) }
-        .recoverWith { case _ => Future.failed(new RuntimeException(s"invalid channel identifier '$channelIdentifier'")) }
-      res <- appKit.register ? fwdReq
-    } yield res
+
 }
